requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django
idna==2.8
pycryptodome==3.9.3
pytz==2019.3
requests==2.22.0
simplejson==3.16.0
sqlparse==0.3.0
urllib3==1.25.7
